% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnpj.R
\name{cnpj}
\alias{cnpj}
\title{Funtion cnpj}
\usage{
cnpj(cnpj)
}
\arguments{
\item{vector}{with Brazilian companies tax numbers}
}
\value{
A data.frame with companies' information
}
\description{
This function returns information about Brazilian companies based on tax number
}
\keyword{cnpj,}
\keyword{number}
\keyword{tax}
