% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_br.R
\name{number_br}
\alias{number_br}
\title{Função number_br}
\usage{
number_br(str)
}
\arguments{
\item{vector}{of numbers in format of character strings}
}
\value{
numeric vector
}
\description{
This converts numbers to English format.
}
\keyword{numbers}
