% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postal.R
\name{postal}
\alias{postal}
\title{Função postal}
\usage{
postal(cep)
}
\arguments{
\item{vector}{with Brazilian postal codes}
}
\value{
A data.frame with the addresses
}
\description{
This function return address based on Brazilian postal code
}
\keyword{cep}
\keyword{code,}
\keyword{postal}
