% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/via_cep.R
\name{via_cep}
\alias{via_cep}
\title{Function via_cep}
\usage{
via_cep(cep, latlong = TRUE)
}
\arguments{
\item{cep}{vector with Brazilian postal codes}

\item{latlong}{boolean. Default=TRUE, it also returns the latlong}
}
\value{
A data.frame with the addresses and latlongs
}
\description{
This function returns address and latlong based on Brazilian postal code
}
\examples{
via_cep(02930000)
via_cep("02929020")
via_cep("02846-000")
}
\keyword{cep,}
\keyword{code,}
\keyword{latlong}
\keyword{via_cep}
