% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtsumR.R
\name{xtsumR}
\alias{xtsumR}
\title{Function xtsumR}
\usage{
xtsumR(df, columns, individuals)
}
\arguments{
\item{df}{data.frame}

\item{columns}{vector with selected colums' numbers}

\item{individuals}{column where the individuals are}
}
\value{
A data.frame with the summary
}
\description{
This function does the same as the xtsum funtion in Stata.
}
\examples{
df<-data.frame(a=rnorm(n=20,mean=23,sd=6),
    b=rnorm(n=20,mean=18,sd=7),
    uf=sample(c("SP","MS","RS","AM","SE"),20,replace=TRUE))
xtsumR(df,c(1,2),"uf")
}
\keyword{data}
\keyword{panel}
\keyword{summary,}
