% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnpj.R
\name{cnpj}
\alias{cnpj}
\title{Function cnpj}
\usage{
cnpj(cnpj, latlong = TRUE)
}
\arguments{
\item{cnpj}{vector with Brazilian companies tax numbers}

\item{latlong}{boolean. Default TRUE, it also returns the latlong}
}
\value{
A data.frame with companies' information
}
\description{
This function returns information about Brazilian companies based on tax number
}
\examples{
cnpj("60.746.948.0001-12")
cnpj("60746948000112")
}
\keyword{address,}
\keyword{cnpj,}
\keyword{latlong}
\keyword{number,}
\keyword{tax}
