% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_br.R
\name{number_br}
\alias{number_br}
\title{Function number_br}
\usage{
number_br(str)
}
\arguments{
\item{str}{vector of numbers in character strings's format}
}
\value{
numeric vector
}
\description{
Converts strings to numbers to English format.
}
\keyword{numbers}
