% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtsumR.R
\name{xtsumR}
\alias{xtsumR}
\title{Função xtsumR}
\usage{
xtsumR(df, columns, individuals)
}
\arguments{
\item{df}{data frame}

\item{columns}{vector with selected colums' numbers}

\item{individuals}{column where the individuals are}
}
\value{
A data.frame with the summary
}
\description{
This function does the same as the xtsum funtion in Stata.
}
\examples{
xtsum(df,c(3,4,5),"uf")
}
\keyword{data}
\keyword{panel}
\keyword{summary,}
